Changed variable font technique

+++ src/_includes/filters/formatpostdate.js
@@ -1,10 +1,13 @@
 /*
- * Format the date of a post into Month Day, Year
+ * Format the date of a post into Month Day, Year (in US English)
  *
- * Input contexts include:
+ * Date input is from the frontmatter of each post, in the form of yyyy-mm-dd
+ *
+ * Usage contexts include:
  *  - index page
  *  - archives page
  *  - individual post pages
+ *
  *
  * Output:
  * - date string in the form of "September 1, 2020"
@@ -13,7 +16,7 @@
 
 module.exports = function (dateString) {
   return dateString.toLocaleString("en-US", {
-    month: "long",
+    month: "short",
     day: "numeric",
     year: "numeric",
   });
